break out browser and helper methods
