+++ src/shared/components/app/navbar.tsx
@@ -11,14 +11,14 @@ import { UserService } from "../../services";
 import { HttpService, RequestState } from "../../services/HttpService";
 import {
   donateLemmyUrl,
-  isBrowser,
   myAuth,
   numToSI,
-  poll,
   showAvatars,
   toast,
   updateUnreadCountsInterval,
 } from "../../utils";
+import { isBrowser } from "../../utils/browser/is-browser";
+import { poll } from "../../utils/helpers/poll";
 import { amAdmin } from "../../utils/roles/am-admin";
 import { canCreateCommunity } from "../../utils/roles/can-create-community";
 import { Icon } from "../common/icon";
