+++ src/shared/utils/browser/is-browser.ts
@@ -1,0 +1,3 @@
+export function isBrowser() {
+  return typeof window !== "undefined";
+}
