+++ src/shared/utils/helpers/sleep.ts
@@ -1,0 +1,3 @@
+export function sleep(millis: number): Promise<void> {
+  return new Promise(resolve => setTimeout(resolve, millis));
+}
