+++ src/shared/utils/types/query-params.ts
@@ -1,0 +1,3 @@
+export type QueryParams<T extends Record<string, any>> = {
+  [key in keyof T]?: string;
+};
